Commit on 06 Jul 2024 - pass float variable in abs function

+++ camastersoftware.com/app/Views/firm_panel/staff_administration/articleship_leave_cal.php
@@ -650,7 +650,7 @@ if (!empty($leaveCalDataArr)) {
         }
 
         var timeDifference = completion_date.getTime() - start_date.getTime();
-        var daysDifference = Math.abs(timeDifference / (1000 * 60 * 60 * 24));
+        var daysDifference = Math.abs((float)timeDifference / (1000 * 60 * 60 * 24));
         if (isNaN(daysDifference)) {
             daysDifference = 0;
         }
